requirements.txt
pycrypto==2.6.1
gevent==1.4.0